Updated description for PhysicalBone.
I've finally figured out what causes most problems with wrong ragdoll behavior, and it's that center of mass is located in Physicalbone and you're free to change it. I've never seen anyone say that, haven't seen it anywhere in the docs and no one was able to help me. But here it is, solution was very simple in the end. 
If you create PhysicalBone and select a bone it is attached too, it automatically moves to the HEAD of a selected bone, making it look like it has to be there, and if you change it's position, it'll be oribiting the bone. Turns out that no, you are free to move it to change it's center of mass. And you HAVE to move it to the center of bone if you're making a ragdoll, otherwise your physbox would have center of mass out of collision shape's bounds and ragdoll would do weird things.